+++ src/game/GameObject.h
@@ -723,7 +723,7 @@ class MANGOS_DLL_SPEC GameObject : public WorldObject
         void AddUse() { ++m_useTimes; }
 
         uint32 GetUseCount() const { return m_useTimes; }
-        uint32 GetUniqueUseCount() const { return m_UniqueUsers.size(); }
+        uint32 GetUniqueUseCount() const { return (uint32)m_UniqueUsers.size(); }
 
         void SaveRespawnTime() override;
