+++ src/game/Unit.h
@@ -2487,7 +2487,7 @@ class MANGOS_DLL_SPEC Unit : public WorldObject
                 m_visibleAuras[slot] = spellid;
         }
         VisibleAuraMap const& GetVisibleAuras() const { return m_visibleAuras; }
-        uint8 GetVisibleAurasCount() const { return m_visibleAuras.size(); }
+        uint8 GetVisibleAurasCount() const { return uint8(m_visibleAuras.size()); }
 
         Aura* GetAura(uint32 spellId, SpellEffectIndex effindex);
         Aura* GetAura(AuraType type, SpellFamily family, uint64 familyFlag, uint32 familyFlag2 = 0, ObjectGuid casterGuid = ObjectGuid());
